Git: Refactor git functions into separate file
